Add accumulator as a demo
Add accumulator.js as a demo of a consumer embedded in the server
that accumulates deltas by path, newer deltas replacing older ones, and
then periodically constructs an array of deltas and prints it to stdout.

If you were interested in reporting all or some data points to a cloud
service this is how I'd go about it.

If there's interest I can pull this in to master with a way to add your own
handler instead of the console.log and configurable period.