Fix ENOENT if bower_components is missing
Currently, if after installing signalk-server-node - but before running
any of the `bower install` commands - you open `localhost:3000` in a
browser window, you'll see a rather uninviting error. This makes sure
that instead you see the rather uninviting "Signal K Server" empty page
instead.

Also, a minor edit to the regex in lib/interfaces/index.js which
requires and exports files in that directory. It only checked for the
existence of .js in filenames, now it checks for files ending in .js.
This is important if you use an editor like vim which creates .swp
files.