Fix path handling in cli.js
This corrects several issues in cli.js where paths were not handled
properly.

Also, adding a symlink to volare-file-settings.json to make the default
case of running bin/signalk-server work correctly. This is not ideal,
it would be much better to have a default settings file which didn't
actually do anything.

However, with this commit it is possible to use pm2 to run the server
as a service (and without elevated permissions).

TODO: Configure server to drop permissions if started as root