fix: always consume pipelement output with pipe(devnull) (#172)
Fixes #171. Multiplexedlog is getting stuck after 16 pushed json deltas, because
the internal buffer fills up. I have no idea why this happens with just this
provider and not others.
The deltas are actually handled with .on('data',...), so we can safely pipe
the ouput to devnull. However this is confusing - it would be better to
create a Writable  end element to the pipe and handle deltas finally
there. This way we could remove the explicit  calls
in Multiplexedlog.