fix: handle vessels.self context properly
Previously a delta with context 'vessels.self' created a vessel with id
self, which confusingly had nothing to do with response to http requests
with /vessels/self path, which fetch data from under the vessel
identified with the self id from the configuration.

This fixes the situation so that deltas with 'vessels.self' context are
routed to under the vessel with the self id.