fix: fix serialport Buffer handling
Adding support for timestamped NMEA0183 playback introduced a bug when
using data from serialport. Serialport has an internal line parser
and emits lines as Buffers, which NMEA0183 to Signal K parser converts
to string as part of the parsing process. To detect that we are playing
back timestamped data we need to check more than just input being
an Object.