fix: tcp port in /signalk and mdns
Fix the default port for deltas over tcp to be 3858. Replace the hard coded
value in /signalk response with the one actually used and omit the tcp
service if it is not active. Fix mdns to use the actual port in use
in the service instead of the http/ws port.