fix: n2k file playback throttling, naming
An n2k file playback configured from the ui needs
to have timestamp throttle, so pass the whole
options object down in pipeStartByType so that the
input type is available when constructing the end of
the pipeline.

style: rename classes with capital first letter

Common JavaScript naming practice, makes it more
obvious what is a function and what is a class.