fix: reorganise nmea0183 passing
Simple provider created from the admin ui was not
passing nmea0183 events at all, because the logic was in
pipedprovider pipeline construction. Providers can send
events directly because they have access to the app
and app.signalk, so use that to send nmea0183 events
from the provider and remove the related logic
from pipedprovider pipeline construction.