fix: app.savePluginOptions from plugin
The enabled and enableLogging flags that are managed
by the server and not part of the plugin's config
schema were not handled when calling savePluginOptions
from a plugin. This changes the logic
so that if a plugin calls savePluginOptions the
values for these two properties are retrieved from
the file system and the plugin config is saved
in the proper structure, with the plugin's own data
nested under 'configuration' property.