fix: handle serial port buffer overrun
Previously there was no limit to how much data would
be written to serialport, which would result in ever growing
buffers and eventual server freeze if the output rate was
more than the serialport could write.

Now there is option maxPendingWrites, which specifies a high
water mark after which input is simply ignored until
the outgoing buffer is below the mark.

This also adds a debug statement to debug output and
possible overruns.