feature: start with tcp ports in use
People seem to be frequently starting the server manually
with the systemd managed one already started and the startup
fails with enigmatic EADDRINUSE 10110 stuff. This changes the
server so that it just logs error about the overlapping
NMEA0183 and Signal K server tcp ports and proceeds to
start the server.