test: add test for running as included module
This adds a separate test for checking that the server can
start up when included as a module. The needed package.json and
the test itself are copied into a temporary directory, then
the server install is npm pack'd and installed into the temp
directory. Everything I tried within the server directory
ended up with npm complaining or shortcutting things so
that the test either did not work or tested nothing.