fix: find admin ui when embedded
If you install signalk-server as a dependency with
npm install --save signalk-server
the admin ui ends up installed rooted at the same
level as signalk-server and the server fails to start, unless
we look for the admin ui also there.

This fix is not complete, as the admin ui will not work
like this, but the server starts.