chore: remove Node version check
Prior to Node 6 or something people frequently tried to run with
extremely old versions of Node, which resulted in not very useful
error messages. Now that we are moving to 10 in #830, but everything still
works on 8, we don't want to force upgrading and create a support
storm. So let's drop the version requirement and see what the response is.