feature: wait for interfaces to start
We don't want providers to start before interfaces, especially plugins,
are started and are ready to handle incoming data.

With this change interfaces and individual plugins can return
a Promise that must be resolved when it has started.

Any failing Promise will interrupt the server's startup - this is
probably not what we want?