feature: disable server update when in Docker
A server running inside docker is not really updateable, as one
should pull a new image, not try to run update inside the container.
Which will not work anyway, because inside the container the
installation is not from npm.

So disable update from the admin ui and show an extra message
for Docker users. This will work with only the official images
with IS_IN_DOCKER set to true, but if you build your own image
you're on your own anyway.