feature: plugin example in ts, typings
Add Plugin example in TypeScript, improve typings.

The changes in plugins typings are NOT backwards compatible, but
as this is a really really new feature and actually does not
change the plugin API in any way, just the typings, I don't
think we really need to care about it yet.