feature: optimise rendering
Use PureComponent so that we don't need to rerender
the pretty heavy html display components but once for
each row.

Create a running, unique key for each row so that React does
not need to modify DOM more than removing the first entry and
adding the last one.

At least that's how this is supposed to work, and there
is a huge difference in performance.