feature: add null removal and ignored sentences to 0183 input (#1032)
* feature: add replacer pipeelement

Add regexp based replacer, that can be used as a pipedprovider
element after liner. Pass in options that have a property
replacers that is an array of {regexp, template} objects.
All the replacers are run on a line in sequence, like
line.replace(new RegExp(regexp, 'gu'), template).

* feature: add null removal and ignored 0183 sentences

Add options to ignore null characters from the 0183 input and
configurable 0183 sentences to remove from the input stream.

These use the generic Replacer pipe element.