feature: add support for zones
Add support for retrieving the zones metadata through the REST API.
Add support for HTTP requests ending in .../meta/zones. Update .../meta to include the zones as well.
As per the spec, if no zones are defined for a requested path, the zones key will have a value of [].
After the change, the zones for all the paths can be defined in defaults.json according to the signalk data model.