feature: debug all server events
This adds the ability to debug log all server 'app'
events, useful to see events' data in the server log.
The eventDebugs debug function cache could be used to provide
a popup menu in the UI to select events that the user
wants to see in the server log.

The event debugs are named signalk-server:events:<eventname>