feature: add global error handler
Now uncaught exceptions crash the server, which makes
debugging things a bit hard, as the server may be crashing
so that admin ui is not available and the server log that
way even less so.

THis adds a global error handler that simply logs the error.
This should provide easier access to errors in the server
log and keep the server up, even if for example a misbehaving
plugin is not handling its own errors.