feature: add setPluginStatus/Error and deprecate setProvider versions
Start actively deprecating 'Provider' concept, starting with plugins, that
never were providers in the first place.

Logs usage of either of the old versions once with a url that points to the
documentation. I tried to craft the message to be neutral enough so that
we don't get support requests from end users about this.

No idea if we can ever really deprecate these without having all plugins that
have not been updated for ages also deprecated, but it's a start.