feature: allow ssl to be terminated on a load balancer
If the SSL connection is terminated on a load balancer then SK is not
aware that clients should use https. This change allows SK to look for
the x-forwarded-proto header so that it knows the connection should use
https.

For this to work the load balancer must provide the x-forwarded-proto
header.

As an example for using SK in a container with traefik on the front end,
the docker-compose file would require the following tags for the sk container:

labels:
  - traefik.enable=true
  # redirect to HTTPS
  - traefik.http.routers.xprotoredirect.rule=Host(`sk.mydomain.com`) && Headers(`X-Forwarded-Proto`, `https`)
  - traefik.http.routers.xprotoredirect.entrypoints=web
  - traefik.http.routers.xprotoredirect.middlewares=redirect@docker
  # HTTPS
  - traefik.http.routers.signalk-prod-web-secure.rule=Host(`sk.mydomain.com`)
  - traefik.http.routers.signalk-prod-web-secure.tls.certresolver=mydnschallenge
  - traefik.http.routers.signalk-prod-web-secure.tls=true
  - traefik.http.routers.signalk-prod-web-secure.entrypoints=websecure