fix: scroll properly to connection form's top
Several people have missed the Data Type selection, because the connection
form is scrolled to the top of the window, not to the top of the visible area.

This hardcodes the header's height as scroll-margin-top so that the top of
the connection form is visible and we don't look like we have support only for
NMEA 2000.