fix: security settings with direct url (#1178)
Not acting on componentDidUpdate caused these pages to show
incorrect / no data when used with a direct url, like when you
reload the page in server.

A better fix would be to have this data in Redux and not handled
independently on the page itself. Sending updates via serverevents
would allow all open admin pages to update and not require a reload.