fix: safe plugin configuration object
Now that plugin enable is outside the plugin configuration form and
not part of it the user may enable the plugin without touching the
configuration. Without this check the plugin gets passed undefined
as configuration, that most plugins probably won't expect and will
fail. The failures should not affect the server's base functions,
but the log messages can be cryptic.