refactor: modify user to be optionally undefined
The fix in previous commit should have been caught by types, but user
was always defined.

This commit changes user to be optionally null, that induces the
change in the order of the arguments, as optionally undefined
arguments can not be before mandatory arguments.

This unearths the mismatch between ContextMatcher and ContextFilter,
that seem to be in conflict?