feature: use debug wrapper
The internal structure of debug modules changed in 4.3.0
https://github.com/debug-js/debug/pull/740 so debug no longer
maintains an internal list of created debug instances.
This means that we can no longer get a list of known
debug instances for the Server Log admin UI page.

This changes the server code so that all code that uses
the added createDebug call get their debug key included
in the list.

I think the change also means that the settings in Server
Log no longer take effect on older debug instances that
are transitive dependencies.