fix: docker build (#1377)
Now that we are using npm workspaces npm install does
not install the packages from npmjs.org, but relies on
linking. Any packages that need explicit building need
to be built. I did not check if adding a postinstall
script would have worked, but explicit building seems
to fix things.

This commit adds also express as a explicit devDependency.
The build step needs this explicitly.