feature: add openapi Course API to plugin API
This commit adds the code needed for generating
a TypeScript client for the Course API and adds
the client to the App object that is passed to
plugin constructors so that plugins have access
to the Course API.

This also changes server-api transpile target to
es2021, because the generated code uses private
properties. This naturally requires a fresh Node
version that won't fly with Node 10.

TypeScript is also updated to the latest version
and the so far unused typedoc is dropped, we can
bring it back once it is actually in use.

Open issues:
- hard coded server address
- the generated code has an import for
  AbortController that I don't know what to do
  with. So far I just commented it out. Needs
  to be solved, prepublishOnly won't work like
  like this
- node-fetch is only an esm module starting in
  version 3. Could not find a way to deal with
  that, so pinned it to v2