feature: add client
This adds client$source property to WebSocket hello message and uses that
same value as $source value for all updates in deltas received from that
WebSocket that do not have source object or $source property already
when received from the client.

The purpose is to create a mechanism for a ws client to tell which of
the messages were originally sent by itself.