fix: empty path subscription period handling
Period and minPeriod do not quite make sense with
empty path subscriptions: for different values the
path is the same, so how do you send data only every
x milliseconds, as it means dropping some of the
values and not sending them at all.

This changes subscription handling so that period
and minPeriod are silently ignored for subscriptions
with an empty path.