chore: build js just once, install server with -g
Build all the server related js code only once and store them as
artifacts. Modify the server's dependencies by removing workspaces
configuration and using npm i from each tarball to update the versions
to point to the tarballs.

Then in reach platform specific Docker build use the prebuilt server
and the modules from artifacts.

This PR also changes the master build to be more in line with the
release build so that the server is installed globally with -g.