refactor: course types fiddling
Restructure the course types to have distinct types instead of
a single PointDestination with all properties optional, that does
not really document the structure of the different input types.

Drop PointDestination.type, as it was not used anywhere.

Instead add CoursePointType to be used as the CourseInfo
nextPoint/previousPoint type discriminator.