fix: multiplexed file detection for plugin output
Logged output from plugins has the plugin id in the
discriminator field instead of the one letter data type
discriminator. If the first line of the multiplexed file
is from a plugin the file type detection does not work
and the file fails to play back silently.

Assert the minimum number of ; delimiters to be 3
instead of checking that charAt(15) is ;.