fix: guard against missing plugin schema
Loading the configuration form for a plugin with no schema
would crash the configuration ui and the browser would
persist the configuration ui state so that subsequent
attempts to navigate to plugin configuration would try
to open the problematic plugin's configuration, again
resulting in a ui crash.

Change the server to always provide an empty schema in
case the plugin is not defining one and log an error
message about it when the plugin configuration page
is loaded.