Refactor UI of the Appstore page
User wants to browser and install plugins faster. User feedback
mentioned bad experience and confusin filters and search on the appstore
page.
A new AppStore and more modern page has been designed.
Easier and less confusing plugin/webapp search is made with a refactor
of the filters and the implementation of a new grid component.
This constitute a boiler plate for future development on the appstore.
The usage of ag-grid improves the user and developer experience to
develop new features on top. The library offers great built-in functionality such as cell renderers columns sorting,
filtering, quick search etc...

Fortawesome was installed using package json
as documented on fontawesome website.
Work in progress loose end further implementation are required.

Resolves: #1675