Created new column for Version number on appStore page:
A new column for the version number on the appstore page.
A few users complained about getting confused and lost in their plugin
search because of version number in between plugin name.

Created a dedicated column for the version number. Also removed the last
updated date for a better display.
Added media query for the header in order to improve display on mobile
version of the search bar.