fix: don't subscribe to notifications in sk connections
@signalk/client subscribes by default to notifications.*.
This is confusing in the first place if the user enters an
explicit subscription and gets more than they subscribed.

Furthermore this wreaks havoc if you have two interconnected
SK servers. Their explicit subscriptions will not create a
loop if they don't touch the same paths, but the default
notifications subscription will if there is even one
active notification.