feature: Add base delta config for anchor roller height
When calculating scope for anchoring, distance of the anchor roller
(or bridle connection point) above the water line should be added
to the depth before applying whatever formula or multiplier the
sailor is using. Like beam and draft, it's a fixed quality of
the vessel, so isn't going to come from a sensor. Having it as a
base delta would allow deriving an "anchoring depth" value from
draft and depthBelowSurface.