feature: default providers for resource types
Not all resource providers will support creating / updating resources
_e.g. `charts-plugin`_.

When there are multiple providers registered for a resource type, the
first plugin / provider registered will be used for write operations
unless the provider to use is specified in the request.

**Definition:** `default provider` = provider used for write operations
_(when `?provider=<plugin-id>` is not specified in the request)_.

Provide the ability to view  and set the `default provider` for specific
resource types and persist the selection bewteen restarts.

Implemented in this PR:
1. GET `/resources/{resourcetype}/_providers` endpoint -> List the
registered resource providers for the resource type
2. GET `/resources/{resourcetype}/_default` endpoint -> Return the
default provider for the resource type
3. POST `/resources/{resourcetype}/_default/{providerid}` -> Set the
default provider for a resource type

_Example:_ Set the default provider for a routes to plugin with the id
`my-provider-id`
```
POST "/resources/routes/_default/my-provider-id"
```