feature: add client side cookie to expose login info
Webapps are often used in multiple windows and the browser's login
status in session cookie may change in another window or in another
webapp (Admin UI vs Freeboard, Kip etc). The regular session cookie
is set to be httpOnly, so it is not available for clientside code.

This adds a separate cookie that is set and removed by the server
when the user's login status changes. A webapp can poll for changes
of this cookie to react. One notable use case for this is
applicationData that is available only when the user/browser is
logged in.