feature: redirect to webapp by disiplayName (#1822)
Signal K webapps have displayName property. This PR adds automatic redirect
to that webapp when the server's root is accessed with a hostname whose
first element is the webapp's displayName.

So if you have a webapp whose displayName is foo and you access the server
using url like http://foo.bar.org:3000/ you will be redirected to that
webapp instead of the default landingPage that is the Admin webapp.