fix: make serialport module optional
serialport module has been optional, as it is platform specific and
its install may fail but otherwise the server would be in working order.

However when converting serialports.js to ts using import had made
it a hard dependency - the server would not start without it,
even if npm install had completed without errors but without
serialport still installed.

Use dynamic require to load it so that even if the module
is missing the server starts.