feat: Add ESM support for plugins
This adds the ability to develop plugins using ESM (`export default`) in addition to the existing CJS (`module.exports =`) support. ESM is widely considered the module format of the future and has been well-supported in NodeJS for many years.

For now, the server first attempts to load a plugin using `require` (CJS), and if that fails, it will then attempt to load with `import` (ESM). Eventually it may make sense to switch this order.