chore: build release with --workspace
This commit changes the release workflow to build all the packages and
the server using npm workspaces. Packages are then published when we
have a newer version than in Github. The motivation for the change
is that the monorepo was recently structure to use shared configuration
and dev dependencies and it is no longer feasible to build the
packages separately.

Previous independent jobs are now steps in one build_and_publish job.