chore: build release with --workspace (#2007)
* chore: build release with --workspace

This commit changes the release workflow to build all the packages and
the server using npm workspaces. Packages are then published when we
have a newer version than in Github. The motivation for the change
is that the monorepo was recently structure to use shared configuration
and dev dependencies and it is no longer feasible to build the
packages separately.

Remove prepublishOnly: GH Action will take care of building, no need to build one
more time. Linting also unnecessary at this stage.