fix: embedded webapp icons not showing (#2079)
* fix: embedded webapp icons not showing

Move all logic related to Webapp card rendering to
Webapp.js, passing complete webAppInfo as prop.

Fix embedded webapp icons not showing by generating
the bgImage url in the same way for all webapps
instead of using the webapp's url as the base, as
it is different for embedded webapps.

* refactor: remove unused properties and related logic