fix: installed version for betas
Previously the installed version would show correctly only when there is
an update available. This makes sense for normal versions, because if
there is no update available you are running the latest version.

However, for beta versions there probably is no update yet available, as
you are have a prerelease version.

This comment changes the logic to always show installed version, if that
is available.