remove fluff
The best practises list is so random and
generic that i'd like to not include it here.

No need to reiterate what your plugin does
here, people can see the code and that
it implements delta emission. In mem
cache is an implementation detail.

No need to tell that the example
referred is good - if we reference
it that is promotion enough, I think.